refactor(desktop): split TopToolbar god-component into menus and hooks (#350)
