+++ apps/geolibre-desktop/src/components/layout/toolbar/ProjectFileDialogs.tsx
@@ -1,0 +1,124 @@
+import {
+  Button,
+  Dialog,
+  DialogContent,
+  DialogDescription,
+  DialogHeader,
+  DialogTitle,
+  Input,
+  Label,
+} from "@geolibre/ui";
+import { useTranslation } from "react-i18next";
+import type { useProjectFileActions } from "../../../hooks/useProjectFileActions";
+
+interface ProjectFileDialogsProps {
+  projectFiles: ReturnType<typeof useProjectFileActions>;
+}
+
+/** The project-file dialogs: Open-from-URL, the error dialog, and the env-var strip prompt. */
+export function ProjectFileDialogs({ projectFiles }: ProjectFileDialogsProps) {
+  const { t } = useTranslation();
+
+  return (
+    <>
+      <Dialog
+        open={projectFiles.projectUrlDialogOpen}
+        onOpenChange={projectFiles.handleProjectUrlDialogOpenChange}
+      >
+        <DialogContent className="max-w-lg">
+          <DialogHeader>
+            <DialogTitle>{t("toolbar.item.openProjectFromUrl")}</DialogTitle>
+            <DialogDescription>
+              {t("toolbar.item.openProjectFromUrlDesc")}
+            </DialogDescription>
+          </DialogHeader>
+          <form className="space-y-4" onSubmit={projectFiles.handleOpenFromUrl}>
+            <div className="space-y-2">
+              <Label htmlFor="project-url">{t("toolbar.item.projectUrl")}</Label>
+              <Input
+                id="project-url"
+                placeholder="https://example.com/project.geolibre.json"
+                value={projectFiles.projectUrl}
+                onChange={(event) => {
+                  projectFiles.setProjectUrl(event.target.value);
+                  projectFiles.setProjectUrlError(null);
+                }}
+              />
+              {projectFiles.projectUrlError ? (
+                <p className="text-xs text-destructive">
+                  {projectFiles.projectUrlError}
+                </p>
+              ) : null}
+            </div>
+            <div className="flex justify-end gap-2">
+              <Button
+                type="button"
+                variant="outline"
+                onClick={() => projectFiles.setProjectUrlDialogOpen(false)}
+              >
+                {t("common.cancel")}
+              </Button>
+              <Button type="submit" disabled={projectFiles.projectUrlLoading}>
+                {projectFiles.projectUrlLoading
+                  ? t("toolbar.item.opening")
+                  : t("toolbar.item.open")}
+              </Button>
+            </div>
+          </form>
+        </DialogContent>
+      </Dialog>
+      <Dialog
+        open={projectFiles.actionError !== null}
+        onOpenChange={(open: boolean) => {
+          if (!open) projectFiles.setActionError(null);
+        }}
+      >
+        <DialogContent className="max-w-lg">
+          <DialogHeader>
+            <DialogTitle>{t("toolbar.item.somethingWentWrong")}</DialogTitle>
+            <DialogDescription>{projectFiles.actionError}</DialogDescription>
+          </DialogHeader>
+          <div className="flex justify-end">
+            <Button onClick={() => projectFiles.setActionError(null)}>
+              {t("toolbar.item.dismiss")}
+            </Button>
+          </div>
+        </DialogContent>
+      </Dialog>
+      <Dialog
+        open={projectFiles.envStripPrompt !== null}
+        onOpenChange={(open: boolean) => {
+          if (!open) projectFiles.resolveEnvStripPrompt("cancel");
+        }}
+      >
+        <DialogContent className="max-w-lg">
+          <DialogHeader>
+            <DialogTitle>{t("settings.env.stripPromptTitle")}</DialogTitle>
+            <DialogDescription>
+              {t("settings.env.stripPromptDesc", {
+                count: projectFiles.envStripPrompt?.count ?? 0,
+              })}
+            </DialogDescription>
+          </DialogHeader>
+          <div className="flex justify-end gap-2">
+            <Button
+              variant="outline"
+              onClick={() => projectFiles.resolveEnvStripPrompt("cancel")}
+            >
+              {t("common.cancel")}
+            </Button>
+            <Button
+              variant="outline"
+              onClick={() => projectFiles.resolveEnvStripPrompt("keep")}
+            >
+              {t("settings.env.keepButton")}
+            </Button>
+            <Button onClick={() => projectFiles.resolveEnvStripPrompt("strip")}>
+              {t("settings.env.stripButton")}
+            </Button>
+          </div>
+        </DialogContent>
+      </Dialog>
+    </>
+  );
+}
